Files for GWDI strain (REMI-seq) upload
the GWDI_mutant_table.xlsx table came from Nicole Grünheit. More detailed descriptions are in a second tab. Some of the information in the table will be used later, for JBrowse integration
The other two are annotation files for the first strain in the table (row 2). All strains should be automatically annotated and inventories created.